Add support for wrap_cb on parsing contexts
This allows us to have a callback called when we pop the stack,
and gives us an opportunity to check the status of the context
object.

Usually we just make sure that we make the final adjustments to the
object, for example, parsing attributes that the user modified and
adjusting the main fields that dictate behaviour.

But it can also be used to generate notices about misconfigured
objects.