Rework action renders for index and select
Move the cleanup and sanity section from the render to a wrap
function. This makes sure that the cleaning up is only done once, and
as soon as the object is created. Thus we can now render the object
multiple times if we want it.

We also took the opportunity to extract the common parts of select and
index rendering, and merge it into a single code path.