Add consistent sort order to pages on our tree
The actual code changed removes the sort from the .select and switches
to a Perl sort block. This allows us to overcome the .select order_by
Limitation of only allowing a single level of sort.

We now have a two-level sort that allows us to implement the following
policy:

* index.md is first;
* then the list of folders, alphabetically sorted;
* then the rest of the files, alphabetically sorted.

Any one of those can override the weight with index_order attribute.